Fix integer overflow
Global tests count was not aligned with db and rest api adjustments.

+++ app/src/main/java/com/mobilecoronatracker/ui/accumulatedcharts/AccumulatedChartsViewModel.kt
@@ -41,7 +41,7 @@ class AccumulatedChartsViewModel(
             AccumulatedChartsViewModelable.GlobalDataWithToday(
                 today.cases, today.todayCases, today.deaths, today.todayDeaths,
                 today.cases - today.deaths - today.recovered, today.recovered,
-                today.recovered - yesterday.recovered, today.critical,  today.tests.toInt()
+                today.recovered - yesterday.recovered, today.critical, today.tests
             )
         )
     }
